Tidied up my code.
I tried to get the two robots to share the same collision through a for loop. I got pretty close at one point but it wasn't working to a satisfactory level. So I've just left it as two separate blocks. Though I have shortened it down a bit and it's still functional.